requirements.txt
fastapi
uvicorn
starkbank
python-dotenv
httpx
apscheduler
faker
